# store: optimized snappy streamed reading