@GiedriusS 2023-06-27
Did a few unscientific tests in a staging env and peaks seem a bit lower with this change:

![Screenshot from 2023-06-27 15-01-35](https://github.com/thanos-io/thanos/assets/2377233/1f0c0c56-a8ba-47c8-9dca-722e055384d4)

I don't have lots of blocks in staging so I can't reproduce this with 
```
sum(container_memory_working_set_bytes{namespace=~"<some namespace>", container!="", pod!=""})
```